Restructure app
- into services/domains and shared components
