+++ app/components/__test__/VerfahrenCounter.test.tsx
@@ -2,7 +2,7 @@
 import {
   getTestTranslations,
   renderWithTestTranslations,
-} from "tests/util/translationsUtil";
+} from "tests/utils/translationsUtil";
 import { describe, expect, it } from "vitest";
 import { VerfahrenCounter } from "~/components/verfahren/VerfahrenCounter";
