+++ app/components/__test__/VerfahrenLoadMoreButton.test.tsx
@@ -3,7 +3,7 @@ import { MemoryRouter } from "react-router";
 import {
   getTestTranslations,
   renderWithTestTranslations,
-} from "tests/util/translationsUtil";
+} from "tests/utils/translationsUtil";
 import { beforeEach, it, vi } from "vitest";
 import { VerfahrenLoadMoreButton } from "~/components/verfahren/VerfahrenLoadMoreButton";
