+++ app/routes/verfahren.$id.tsx
@@ -3,9 +3,9 @@ import { Await, Link, LoaderFunctionArgs, useLoaderData } from "react-router";
 import Alert from "~/components/Alert";
 import VerfahrenTileSkeleton from "~/components/VerfahrenTileSkeleton.static";
 import VerfahrenTile from "~/components/verfahren/VerfahrenTile";
+import fetchVerfahrenById from "~/domains/verfahren/fetchVerfahrenById.server";
 import { authContext, authMiddleware } from "~/middleware/auth.server";
 import { useTranslations } from "~/services/translations/context";
-import fetchVerfahrenById from "~/services/verfahren/fetchVerfahrenById.server";
 
 // this route requires users to be logged in
 export const middleware = [authMiddleware];
