+++ app/utils/buildSearchParams.ts
