+++ tests/e2e/index.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "@playwright/test";
-import { getTestTranslations } from "tests/util/translationsUtil";
+import { getTestTranslations } from "tests/utils/translationsUtil";
 import { LoginError } from "~/routes/action.login-user";
 import { LogoutType } from "~/routes/action.logout-user";
